+++ HISTORY.rst
@@ -62,4 +62,5 @@ History
 4.0.0 (2020-10-xx)
 ------------------
 * Python 3.7 supported
-* Pagination with continue parameter over linked resources supported
+* Pagination with `continue` parameter over linked resources supported
+* Requirements upgraded