+++ HISTORY.rst
@@ -61,6 +61,7 @@ History
 
 4.0.0 (2020-10-xx)
 ------------------
+* Python 3.5 support removed
 * Python 3.7 supported
 * Python 3.8 supported
 * Pagination with `continue` parameter over linked resources supported