+++ HISTORY.rst
@@ -59,7 +59,7 @@ History
 * Removing `transaction_id` value when the client is created
 * Allowing to provide `user_id` and `application_id` values as correlation id prefix
 
-4.0.0 (2020-10-xx)
+4.0.0 (2020-10-21)
 ------------------
 * Python 3.5 support removed
 * Python 3.7 supported